+++ apps/forms/app/(dev)/dev/templating/page.tsx
@@ -1,8 +1,5 @@
 "use client";
-import { TemplateVariables } from "@/lib/templating";
 import { useEffect, useMemo, useState } from "react";
-import { fmt_local_index } from "@/utils/fmt";
-import { en, ko, Faker } from "@faker-js/faker";
 import { render } from "@/lib/templating/template";
 import {
   Card,
@@ -10,10 +7,6 @@ import {
   CardFooter,
   CardHeader,
 } from "@/components/ui/card";
-import { useEditor, EditorContent } from "@tiptap/react";
-import StarterKit from "@tiptap/starter-kit";
-import Mention from "@tiptap/extension-mention";
-import suggestion from "./suggestion";
 import { z } from "zod";
 import {
   Select,
@@ -29,12 +22,7 @@ import {
   CollapsibleContent,
   CollapsibleTrigger,
 } from "@/components/ui/collapsible";
-import {
-  ArrowRightIcon,
-  CaretLeftIcon,
-  CodeIcon,
-  ReloadIcon,
-} from "@radix-ui/react-icons";
+import { CaretLeftIcon, ReloadIcon } from "@radix-ui/react-icons";
 import { Badge } from "@/components/ui/badge";
 import {
   HoverCard,
@@ -42,60 +30,34 @@ import {
   HoverCardContent,
 } from "@/components/ui/hover-card";
 import { Tabs, TabsContent, TabsList, TabsTrigger } from "@/components/ui/tabs";
-import {
-  Table,
-  TableBody,
-  TableCell,
-  TableHead,
-  TableHeader,
-  TableRow,
-} from "@/components/ui/table";
-import { unescape } from "querystring";
 import resources from "@/k/i18n";
-
-const fakerlocales = {
-  en: en,
-  ko: ko,
-};
-
-function fake(lang: string) {
-  const faker = new Faker({
-    // @ts-ignore
-    locale: fakerlocales[lang],
-  });
-
-  const response_index = faker.number.int(50);
-  const context: TemplateVariables.FormResponseContext = {
-    customer: {
-      short_id: faker.string.uuid(),
-    },
-    form_title: "Buy me a coffee",
-    title: "Buy me a coffee",
-    fields: {},
-    language: "en",
-    response: {
-      index: response_index,
-      idx: fmt_local_index(response_index),
-    },
-    session: {},
-  };
-
-  return context;
-}
+import { useMockedContext } from "@/scaffolds/template-editor/mock";
+import { TemplateTextEditor } from "@/scaffolds/template-editor/text-editor";
+import { ContextVariablesTable } from "@/scaffolds/template-editor/about-variable-table";
 
 export default function TemplatingDevPage() {
-  const [_context_refresh_key, setContextRefreshKey] = useState(0);
+  const [isModified, setIsModified] = useState(false);
+  const [contextRefreshKey, setContextRefreshKey] = useState(0);
   const [texts, setTexts] = useState<Record<string, string>>({
     h1: ExamplePropsZ.shape.h1._def.defaultValue(),
     h2: ExamplePropsZ.shape.h2._def.defaultValue(),
     p: ExamplePropsZ.shape.p._def.defaultValue(),
   });
 
-  const context = useMemo(() => fake("ko"), [_context_refresh_key]);
-
   const onTextChange = (key: string, value: string) => {
     setTexts((prev) => ({ ...prev, [key]: value }));
+    setIsModified(true);
   };
+
+  const context = useMockedContext(
+    {
+      title: "Form Title",
+    },
+    {
+      refreshKey: String(contextRefreshKey),
+      lang: "en",
+    }
+  );
 
   const onReloadContextClick = () => {
     setContextRefreshKey((prev) => prev + 1);
@@ -139,7 +101,15 @@ export default function TemplatingDevPage() {
           </div>
           <div className="flex-1 flex justify-end gap-2">
             <Button variant="secondary">Cancel</Button>
-            <Button>Save</Button>
+            <Button
+              disabled={!isModified}
+              onClick={() => {
+                console.log("save");
+                setIsModified(false);
+              }}
+            >
+              Save
+            </Button>
           </div>
         </div>
       </header>
@@ -149,7 +119,8 @@ export default function TemplatingDevPage() {
         {/*  */}
         <div className="flex-1 h-full p-4">
           <div className="h-full flex items-center justify-center">
-            <Example h1={out.h1} h2={out.h2} p={out.p} />
+            {/* @ts-ignore */}
+            <Example {...out} />
           </div>
         </div>
         <aside className="border-s flex-1 h-full max-w-md overflow-y-scroll">
@@ -245,61 +216,6 @@ export default function TemplatingDevPage() {
   );
 }
 
-interface AboutVariable {
-  key: string;
-  description?: string;
-  optional: boolean;
-}
-
-function flattenSchema(
-  schema: any,
-  parentKey = "",
-  result: AboutVariable[] = []
-) {
-  Object.keys(schema.shape).forEach((key) => {
-    const fullPath = parentKey ? `${parentKey}.${key}` : key;
-    const field = schema.shape[key];
-
-    if (field instanceof z.ZodObject) {
-      flattenSchema(field, fullPath, result);
-    } else {
-      const description = field._def?.description || "";
-      const optional = !!field.isOptional();
-      result.push({ key: fullPath, description, optional });
-    }
-  });
-
-  return result;
-}
-function ContextVariablesTable() {
-  const about = flattenSchema(TemplateVariables.schema);
-
-  return (
-    <Table>
-      <TableHeader>
-        <TableRow>
-          <TableHead className="w-[100px]">key</TableHead>
-          <TableHead>description</TableHead>
-          <TableHead className="text-right">available</TableHead>
-        </TableRow>
-      </TableHeader>
-      <TableBody>
-        {about.map((row) => (
-          <TableRow key={row.key}>
-            <TableCell className="font-medium">
-              <code>{row.key}</code>
-            </TableCell>
-            <TableCell className="text-xs">{row.description}</TableCell>
-            <TableCell className="text-right">
-              {row.optional ? "depends" : "always"}
-            </TableCell>
-          </TableRow>
-        ))}
-      </TableBody>
-    </Table>
-  );
-}
-
 const ExamplePropsZ = z.object({
   h1: z.string().default("{{response.idx}}"),
   h2: z
@@ -337,40 +253,4 @@ function Example({ h1, h2, p }: { h1: string; h2: string; p: string }) {
       {/* <CardFooter className="flex w-full p-0"></CardFooter> */}
     </Card>
   );
-}
-
-function TemplateTextEditor({
-  id,
-  defaultValue,
-  onValueChange,
-}: {
-  id?: string;
-  defaultValue?: string;
-  onValueChange?: (html: string) => void;
-}) {
-  const editor = useEditor({
-    autofocus: true,
-    extensions: [
-      StarterKit.configure({
-        heading: undefined,
-      }),
-      // Mention.configure({
-      //   HTMLAttributes: {
-      //     class: "mention",
-      //   },
-      //   suggestion,
-      // }),
-    ],
-    editorProps: {
-      attributes: {
-        class: "prose dark:prose-invert p-1 focus:outline-none",
-      },
-    },
-    content: defaultValue,
-    onUpdate: ({ editor }) => {
-      onValueChange?.(editor.getHTML());
-    },
-  });
-
-  return <EditorContent id={id} className="h-full w-full" editor={editor} />;
 }
